Update text, intro slide
