+++ app/javascript/dashboard/components-next/year-in-review/slides/ThankYouSlide.vue
@@ -17,15 +17,13 @@ const signatureImage =
 <template>
   <div class="absolute inset-0 flex items-center justify-center px-32 py-20">
     <div class="flex flex-col gap-9 flex-shrink-0 max-w-4xl">
-      <div class="flex gap-14 flex-col">
+      <div class="flex gap-14 flex-col items-start">
         <div
           class="text-3xl md:text-5xl lg:text-6xl font-bold tracking-tight text-black-900"
         >
           {{ t('YEAR_IN_REVIEW.THANK_YOU.TITLE', { year }) }}
         </div>
-        <div
-          class="text-xl lg:text-3xl font-medium tracking-tight text-black-800"
-        >
+        <div class="text-xl lg:text-3xl font-medium text-black-800">
           {{ t('YEAR_IN_REVIEW.THANK_YOU.MESSAGE', { nextYear: year + 1 }) }}
         </div>
         <div>
